requirements.txt
speechrecognition
obs-websocket-py
pyttsx3
requests
PyQt5
pyaudio
